requirements.txt
Django>=4.2,<5.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.1.0

tensorflow==2.18.0
keras==3.8.0
numpy==1.26.0


openai>=1.0.0
python-dotenv